dom/tests/mochitest/bugs/test_bug564208.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=564208
-->
<head>
  <title>Test for Bug 564208</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=564208">Mozilla Bug 564208</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 564208 **/

var sawLength = false;
var c = document.childNodes;
c.__proto__ = document.createTextNode(" ");
for (var i in c) { if (i === "length") sawLength = true; }

ok(sawLength, "property enumerated length");

</script>
</pre>
</body>
</html>
